+++ trs/models.py
@@ -35,8 +35,9 @@ def this_year_week():
     return result
 
 
-def days_missing_per_year():
-    cache_key = 'days_missing_per_year3'
+
+def days_missing_per_year_at_start_and_end():
+    cache_key = 'days_missing_per_year_at_start_and_end1'
     cached = cache.get(cache_key)
     if cached is not None:
         return cached
@@ -44,8 +45,10 @@ def days_missing_per_year():
     from django.conf import settings  # Local import, prevents circular.
     for year in range(settings.TRS_START_YEAR, settings.TRS_END_YEAR + 1):
         first_year_week = YearWeek.objects.filter(year=year).first()
-        missing = first_year_week.first_day.weekday()  # Mon=0, Tue=1, Wed=3
-        result[year] = missing
+        at_start = first_year_week.first_day.weekday()  # Mon=0, Tue=1, Wed=3
+        last_day = datetime.date(year=year, month=12, day=31)
+        at_end = max(0, 4 - last_day.weekday())
+        result[year] = (at_start, at_end)
     cache.set(cache_key, result, 3600 * 24 * 29)  # Max memcache age.
     return result
 
@@ -133,7 +136,7 @@ class Person(models.Model):
         return self.name
 
     def cache_key(self, for_what, year_week=None):
-        cache_version = 1
+        cache_version = 2
         week_id = year_week and year_week.id or this_year_week().id
         return 'person-%s-%s-%s-%s-%s' % (
             self.id, self.cache_indicator, for_what, week_id, cache_version)
@@ -222,8 +225,13 @@ class Person(models.Model):
             if week in changes_per_week:
                 current_amount += changes_per_week[week]
             result += current_amount
-        result -= days_missing_per_year()[this_year] * 8
-        # The line above might have pushed it below zero, so compensate:
+
+        (missing_at_start,
+         missing_at_end) = days_missing_per_year_at_start_and_end()[this_year]
+        result -= missing_at_start * 8
+        if year_week.week >= 52:
+            result -= missing_at_end * 8
+        # The lines above might have pushed it below zero, so compensate:
         return max(0, result)
 
     @cache_on_model
@@ -478,6 +486,9 @@ class FinancialBase(models.Model):
                 # If tls_request doesn't exist we're running tests. Adding
                 # this 'if' is handier than mocking it the whole time :-)
                 self.added_by = tls_request.user
+        self.project.save()
+        # ^^^ Project is available on subclasses. This increments the cache
+        # key.
         return super(FinancialBase, self).save(*args, **kwargs)